backend/requirements.txt
fastapi[standard]
python-dotenv
ultralytics
Pillow
cloudinary
